refactor(example_app): changed layout to support current and future sherpa-onnx features
