+++ example/package.json
@@ -10,8 +10,12 @@
     "build:ios": "react-native build-ios --mode Debug"
   },
   "dependencies": {
+    "@react-navigation/native": "^7.1.28",
+    "@react-navigation/native-stack": "^7.11.0",
     "react": "19.2.0",
-    "react-native": "0.83.0"
+    "react-native": "0.83.0",
+    "react-native-safe-area-context": "^5.6.2",
+    "react-native-screens": "^4.20.0"
   },
   "devDependencies": {
     "@babel/core": "^7.25.2",
