+++ example/src/screens/tts/TTSScreen.tsx
@@ -1,0 +1,72 @@
+import { View, Text, StyleSheet } from 'react-native';
+import { SafeAreaView } from 'react-native-safe-area-context';
+
+export default function TTSScreen() {
+  return (
+    <SafeAreaView style={styles.container}>
+      <View style={styles.content}>
+        <Text style={styles.icon}>🔊</Text>
+        <Text style={styles.title}>Text-to-Speech</Text>
+        <Text style={styles.subtitle}>Coming Soon</Text>
+        <Text style={styles.description}>
+          This feature will allow you to generate speech from text using offline
+          TTS models.
+        </Text>
+        <View style={styles.featureList}>
+          <Text style={styles.featureItem}>• Multiple TTS model support</Text>
+          <Text style={styles.featureItem}>• Voice selection</Text>
+          <Text style={styles.featureItem}>• Speed control</Text>
+          <Text style={styles.featureItem}>• Audio playback</Text>
+        </View>
+      </View>
+    </SafeAreaView>
+  );
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#F2F2F7',
+  },
+  content: {
+    flex: 1,
+    alignItems: 'center',
+    justifyContent: 'center',
+    padding: 32,
+  },
+  icon: {
+    fontSize: 72,
+    marginBottom: 24,
+  },
+  title: {
+    fontSize: 28,
+    fontWeight: 'bold',
+    color: '#000000',
+    marginBottom: 8,
+  },
+  subtitle: {
+    fontSize: 20,
+    fontWeight: '600',
+    color: '#FF9500',
+    marginBottom: 24,
+  },
+  description: {
+    fontSize: 16,
+    color: '#8E8E93',
+    textAlign: 'center',
+    lineHeight: 24,
+    marginBottom: 32,
+  },
+  featureList: {
+    alignSelf: 'stretch',
+    backgroundColor: '#FFFFFF',
+    borderRadius: 12,
+    padding: 20,
+  },
+  featureItem: {
+    fontSize: 15,
+    color: '#000000',
+    marginBottom: 12,
+    lineHeight: 22,
+  },
+});
